requirements.txt
phidata
fastapi
python-dotenv
uvicorn
duckduckgo-search
packaging
groq
yfinance
sqlalchemy